+++ Hw3UnitTests/JaccardSimilarityUnitTests.cs
@@ -14,11 +14,11 @@ namespace Hw3UnitTests
 		public void TestWithVector_ShouldBe0()
 		{
 			var jac = new JaccardSimilarity();
-			double result = jac.CalculateSimilarity(new Article(new Dictionary<int, int>()
+			double result = jac.CalculateSimilarity(new Article(new Dictionary<int, double>()
 			{
 				{0, 1},
 				{1, 2}
-			}), new Article(new Dictionary<int, int>()
+			}), new Article(new Dictionary<int, double>()
 			{
 				{2, 2},
 				{3, 3}
@@ -31,11 +31,11 @@ namespace Hw3UnitTests
 		public void TestWithVector_ShouldBe1()
 		{
 			var jac = new JaccardSimilarity();
-			double result = jac.CalculateSimilarity(new Article(new Dictionary<int, int>()
+			double result = jac.CalculateSimilarity(new Article(new Dictionary<int, double>()
 			{
 				{0, 1},
 				{1, 2}
-			}), new Article(new Dictionary<int, int>()
+			}), new Article(new Dictionary<int, double>()
 			{
 				{0, 1},
 				{1, 2}
@@ -48,11 +48,11 @@ namespace Hw3UnitTests
 		public void TestWithVector_ShouldBe05()
 		{
 			var jac = new JaccardSimilarity();
-			double result = jac.CalculateSimilarity(new Article(new Dictionary<int, int>()
+			double result = jac.CalculateSimilarity(new Article(new Dictionary<int, double>()
 			{
 				{0, 1},
 				{1, 2}
-			}), new Article(new Dictionary<int, int>()
+			}), new Article(new Dictionary<int, double>()
 			{
 				{0, 1},
 				{1, 5}
@@ -65,11 +65,11 @@ namespace Hw3UnitTests
 		public void TestWithVector_ShouldBeDecimal()
 		{
 			var jac = new JaccardSimilarity();
-			double result = jac.CalculateSimilarity(new Article(new Dictionary<int, int>()
+			double result = jac.CalculateSimilarity(new Article(new Dictionary<int, double>()
 			{
 				{0, 1},
 				{1, 2}
-			}), new Article(new Dictionary<int, int>()
+			}), new Article(new Dictionary<int, double>()
 			{
 				{0, 1},
 				{1, 5},
@@ -83,12 +83,12 @@ namespace Hw3UnitTests
 		public void TestWithVector_ShouldBeDecimal2()
 		{
 			var jac = new JaccardSimilarity();
-			double result = jac.CalculateSimilarity(new Article(new Dictionary<int, int>()
+			double result = jac.CalculateSimilarity(new Article(new Dictionary<int, double>()
 			{
 				{0, 1},
 				{1, 2},
 				{3, 10},
-			}), new Article(new Dictionary<int, int>()
+			}), new Article(new Dictionary<int, double>()
 			{
 				{0, 1},
 				{1, 5},
